Move and rename curated-rolls to wishlists
