+++ src/app/item-popup/ItemSockets.tsx
@@ -5,9 +5,9 @@ import React from 'react';
 import './ItemSockets.scss';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
 import { D2Item, DimSocket, DimSocketCategory, DimPlug } from '../inventory/item-types';
-import { InventoryCuratedRoll } from '../curated-rolls/curatedRollService';
+import { InventoryCuratedRoll } from '../wishlists/wishlists';
 import { connect, DispatchProp } from 'react-redux';
-import { wishListsEnabledSelector, inventoryCuratedRollsSelector } from '../curated-rolls/reducer';
+import { wishListsEnabledSelector, inventoryCuratedRollsSelector } from '../wishlists/reducer';
 import { RootState } from '../store/reducers';
 import { getReviews } from '../item-review/reducer';
 import { D2ItemUserReview } from '../item-review/d2-dtr-api-types';
