+++ src/app/search/search-filters.ts
@@ -18,8 +18,8 @@ import { maxLightLoadout } from '../loadout/auto-loadouts';
 import { itemTags, DimItemInfo, getTag, getNotes } from '../inventory/dim-item-info';
 import store from '../store/store';
 import { loadoutsSelector } from '../loadout/reducer';
-import { InventoryCuratedRoll } from '../curated-rolls/curatedRollService';
-import { inventoryCuratedRollsSelector } from '../curated-rolls/reducer';
+import { InventoryCuratedRoll } from '../wishlists/wishlists';
+import { inventoryCuratedRollsSelector } from '../wishlists/reducer';
 import { D2SeasonInfo } from '../inventory/d2-season-info';
 import { D2EventPredicateLookup } from 'data/d2/d2-event-info';
 import { getRating, ratingsSelector, ReviewsState, shouldShowRating } from '../item-review/reducer';
