+++ src/app/wishlists/actions.ts
@@ -1,0 +1,6 @@
+import { createStandardAction } from 'typesafe-actions';
+import { WishList } from './types';
+
+export const loadWishLists = createStandardAction('wishlists/LOAD')<WishList>();
+
+export const clearWishLists = createStandardAction('wishlists/CLEAR')();
