+++ src/app/wishlists/wishlist-file.ts
@@ -1,6 +1,19 @@
-import { CuratedRoll, DimWishList, CuratedRollsAndInfo } from './curatedRoll';
+import { CuratedRoll, DimWishList, WishList } from './types';
 import _ from 'lodash';
-import { getTitle, getDescription } from './curated-roll-metadata-reader';
+
+/* Utilities for reading a wishlist file */
+
+/**
+ * Extracts rolls, title, and description from the meat of
+ * a wish list text file.
+ */
+export function toWishList(fileText: string): WishList {
+  return {
+    curatedRolls: toCuratedRolls(fileText),
+    title: getTitle(fileText),
+    description: getDescription(fileText)
+  };
+}
 
 /** Translate a single banshee-44.com URL -> CuratedRoll. */
 function toCuratedRoll(bansheeTextLine: string): CuratedRoll | null {
@@ -100,14 +113,60 @@ function toCuratedRolls(fileText: string): CuratedRoll[] {
   ).flat();
 }
 
-/**
- * Extracts rolls, title, and description from the meat of
- * a wish list text file.
+function findMatch(sourceFileLine: string, regExToMatch: RegExp): string | undefined {
+  if (!sourceFileLine || !sourceFileLine.length) {
+    return undefined;
+  }
+
+  const matchResults = sourceFileLine.match(regExToMatch);
+
+  if (!matchResults || matchResults.length !== 2) {
+    return undefined;
+  }
+
+  return matchResults[1];
+}
+
+function findTitle(sourceFileLine: string): string | undefined {
+  return findMatch(sourceFileLine, /^title:(.*)/);
+}
+
+function findDescription(sourceFileLine: string): string | undefined {
+  return findMatch(sourceFileLine, /^description:(.*)/);
+}
+
+/*
+ * Will extract the title of a DIM wish list from a source file.
+ * The title should follow the following format:
+ * title:This Is My Source File Title.
+ *
+ * It will only look at the first 20 lines of the file for the title,
+ * and the first line that looks like a title will be returned.
  */
-export function toCuratedRollsAndInfo(fileText: string): CuratedRollsAndInfo {
-  return {
-    curatedRolls: toCuratedRolls(fileText),
-    title: getTitle(fileText),
-    description: getDescription(fileText)
-  };
+function getTitle(sourceFileText: string): string | undefined {
+  if (!sourceFileText) {
+    return undefined;
+  }
+
+  const sourceFileLineArray = sourceFileText.split('\n').slice(0, 20);
+
+  return sourceFileLineArray.map(findTitle).find((s) => s);
+}
+
+/*
+ * Will extract the description of a DIM wish list from a source file.
+ * The description should follow the following format:
+ * description:This Is My Source File Description And Maybe It Is Longer.
+ *
+ * It will only look at the first 20 lines of the file for the description,
+ * and the first line that looks like a description will be returned.
+ */
+function getDescription(sourceFileText: string): string | undefined {
+  if (!sourceFileText) {
+    return undefined;
+  }
+
+  const sourceFileLineArray = sourceFileText.split('\n').slice(0, 20);
+
+  return sourceFileLineArray.map(findDescription).find(Boolean);
 }
